backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.3.2
chromadb==0.4.18
pymupdf==1.23.8
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
celery==5.3.4
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
httpx==0.25.2
tenacity==8.2.3 
numpy==1.26.4 